Set noHostRoute = true for local loopback DSF Ips
Summary:
For local loopback IPs we set noHostRoute to true
since we dont want to install the host/neighbor entry
in HW (the same /128, /32 is covered by ip2me routes).
However we still want to program the encap info (MAC,
encap ID) in HW, so we continue to send the entry to
send it to SDK, but with noHostRoute flag set.

Reviewed By: srikrishnagopu

Differential Revision: D52759880

fbshipit-source-id: 325ff48f5a9c1cf58000d3ecf1eb5dd4e1443338