Send pkts to first interface + loopback IPs
Summary:
On DNX devices we found a bug whereby pkts to
loopback IP hit recycle port and then loop until TTL hits
1 and only then get trapped to (low pri)CPU queue. This
has to do with how we are programming the neighbor entries
for loopback IPs. This was totally missed by our COPP tests.
Enhance tests to start capturing this.

Reviewed By: srikrishnagopu

Differential Revision: D52453221

fbshipit-source-id: 1423fbcc145d4142c33cf548ac43d75359ae50aa