Move phy info update interval from SwSwitch to HwSwitch
Summary: Moving the logic to read phyInfo every 10 seconds to HwSwitch. Reason - Reading it too often is a hardware limitation, hence best if the timestamp checks exist in HwSwitch. Moving it to HwSwitch also makes it easy to handle PhyInfo updates with multi-npu hw agent as in the subsequent diffs

Reviewed By: msomasundaran

Differential Revision:
D52826383

Privacy Context Container: L1125642

fbshipit-source-id: a8a6a262236e1e7ca6f530d478d95630c9f7493d