stop function scheduler without waiting for sw switch state being in fully configured state
Summary:
The diff [fboss] contains code changes in the file fbcode/fboss/agent/SwAgentInitializer.cpp. The changes include removing two lines of code that were causing a deadlock and shutting down the function scheduler if it exists. The summary of the diff is that it removes deadlock-causing code and shuts down the function scheduler if it exists.

This is done to allow stopping function scheduler after stopping SwSwitch. This is needed for SplitSwAgent cold boot shutdown. SplitSwAgent's thrift server must be stopped after stopping SwSwitch as thrift server is needed to send ALPM state as needed for exit for cold boot

Reviewed By: zechengh09

Differential Revision:
D52817924

Privacy Context Container: L1125642

fbshipit-source-id: ad88e7edced7d0263c72f0bf01b80ddcc989f348