wait for async init thread before destroying split sw agent initializer
Summary: This diff contains changes to the FBOSS agent codebase. The changes are made to the SwAgentInitializer.cpp and MultiSwitchAgentEnsemble.cpp files. In SwAgentInitializer.cpp, a new line is added to stop the thrift server before resetting it. In MultiSwitchAgentEnsemble.cpp, a new function is added to join the asyncInitThread before the agentInitializer_ is destroyed. Additionally, a new function is added to AgentEnsemble.h to join the asyncInitThread before the destructor is called. These changes are made to prevent a race condition between stopping the thrift server and destroying the agentInitializer_ object.

Reviewed By: zechengh09

Differential Revision: D52820591

fbshipit-source-id: f6f1f8b15d115aeec518426005e27277b54dfec4