stopAgent method for split SwAgent
Summary:
This diff introduces a new method `stopAgent` in the `SplitSwAgentInitializer` class, which is called when handling the exit signal. The new method stops the agent and reverts to the minimum ALPM state if a file indicating a cold boot is present. Otherwise, it calls the base class's `handleExitSignal` method to gracefully exit for warm boot. The changes are made in the `SplitSwAgentInitializer.h` and `SplitSwAgentInitializer.cpp` files.

The stopAgent method is invoked from AgentTest for exiting Agent for warm boot or cold boot. For cold boot exit, the sequence is

1. Stop SwSwitch, set up minimum ALPM state
2. Stop HwSwitch handler
3. stop thrift server and function scheduler
4. reset initializer

Reviewed By: zechengh09

Differential Revision:
D52818006

Privacy Context Container: L1125642

fbshipit-source-id: 53d5bac76ebaf9d7f39a6759f6e165c4219cc017