Check port type from config
Summary:
Issue:
When agent was cold booted with the port drain state set to 1, we were crashing
here - https://fburl.com/code/vkxjnh37

That is because, orig will not have right portType set. I thought of setting
port type in orig at the called similar to logicalId, but felt checking the
port type from cfg is the right thing to do.

Reviewed By: jasmeetbagga

Differential Revision: D52887269

fbshipit-source-id: 3562c33d933a1231721edf730ccea9fe26011059