AgentHwTest apis to get latest port stats
Summary:
Pull stats from SwSwitch and use them in AgentHwTests.
Few things are a TODO here.
- For mono mode, HwSwitchStats are only getting populated
during FSDB publishing. This is inconsistent with multi-switch
mode and will not work for tests. Will address this in a
subsequent diff.
- We have APIs for getting stats defined in AgentEnsemble as well.
But those will only work for mono mode. We need to retire them
or unify to pull from SwSwitch only.

Reviewed By: zechengh09

Differential Revision: D52932805

fbshipit-source-id: 2a90910bd7df2050cf49be9d798158efc006393b