Populate HwSwitchStats in mono mode as well
Summary:
We had a inconsistency in populating HwSwitchStats.
In that for mono mode we waited till the point of FSDB publish
to populate HwSwitchStats. While for multi mode, we populate
them as we get these via thrift. So for mono mode, populate
as you collect stats from HW

Reviewed By: shri-khare

Differential Revision:
D52937175

Privacy Context Container: L1125642

fbshipit-source-id: f63aaf837b88a3d6c2f73729103dec0acbbe4a11