Parse ethHdr portion of l2 protocols - mka, lldp, lacp
Summary:
We use deep pkt inspection in copp tests for these
protocols. However when doing see, atleast for MKA and LACP
we are actually not constructing the full PDU, but rather
just the eth hdr. Further the pkt structure code for these
is embeded deep into their libs and not abstracted into
pkt lib. So parse up to the eth hdr part only.

Reviewed By: nivinl

Differential Revision: D52960491

fbshipit-source-id: d2c7970dc32d2f34a9c86b63401065a4c12b7fb7