use switch id 20 for fabric switch
Summary:
This diff updates the switch id for fabric switch in the SwitchSettingsTests.cpp and ThriftTest.cpp files. The switch id is changed from 2 to 20 in both files. The TestUtils.cpp file is also updated to reflect the new switch id.

switch id 1 and 2 will be used for voq switches in tests

Reviewed By: jasmeetbagga

Differential Revision:
D52967177

Privacy Context Container: L1125642

fbshipit-source-id: 4301a634aec73047df6099d727568abc17afebe0