update matcher string when switch id is excluded from matcher
Summary:
Defining a new function "getMatcherString" to convert a set of SwitchIDs to a string.

The function is used to create a matcher string for HwSwitchMatcher objects.

Reuse this function in both constructor and in exclude apis to update matcher string when matcher is either constructed or switch id is excluded.

Reviewed By: jasmeetbagga

Differential Revision:
D52967179

Privacy Context Container: L1125642

fbshipit-source-id: 2faf8c23878b48012bf7c647823fc2523590353a