Add new test to validate CPU queue for traffic to unresolved adj route
Summary:
Traffic to route pointing to unresolved adjacency would get punted to CPU and the expectation is
that this traffic goes to CPU queue 0. However, looks like we have an issue on TAJO where this traffic
seems to be hitting queue 2. Adding a new test to cover this scenario, this is an issue in prod now -
[post](https://fb.workplace.com/groups/907812870130827/permalink/1417181599193949/). Raising a case with vendor to address this as looks like there is no explicit programming to
 redirect packets to the right CPU queue in this specific case.

Reviewed By: jasmeetbagga

Differential Revision: D52895721

fbshipit-source-id: ee44618ec965e1fe2fc0bcfd94db383bac225e4d