Export current value of cpu queue counter instead of incremental value
Summary:
cpu queue stats are collected from hwagents and exported to swagent. These stats are monotonic counters. They have the following restrictions
- get gives only increment and not sum
- set sets only absolute value and not increment
The aggregated counter on swswitch was wrong for cpu queue. The fix is to get accumulated value from fb303::fbData and send to swswitch. This is the same mechanism we use for getting acl counters

Reviewed By: phshaikh

Differential Revision: D52929919

fbshipit-source-id: f52dbdaf65bc8fa407d22d0fc3a6727d106012c8