Change the verification logic for nullRouteHit HW test
Summary:
Since inDiscards is not read from HW counters and instead it's derived from
total discards and inDstNullDiscards (drop entry) and pause RX, this may lead to
error in case total discards and inDstNullDiscards are returned out of sync even once
by SDK and even if it gets corrected in subsequent query.

So, excluding inDiscards from the verification logic and instead make sure all drops
in this test account for inDstNullDiscards.

Reviewed By: jasmeetbagga

Differential Revision: D52938847

fbshipit-source-id: d9d8167097df7ad1be0052ed463bcd4cc05738e9