DsfSubscriber: pass per switch systemPort/Intfs to scheduleUpdate
Summary:
FSDB sync carries switchIdTo{systemPort, Intf} info. Previously, DSF subscriber
processed that as a flat list. This worked fine as we had only a single NPU in
an RDSW. As we look to enable multiple NPUs, we need to retain
switchIdTo{systemPort, Intf} info and process it.

This diff: pass that info for processing.

Reviewed By: msomasundaran

Differential Revision: D53167392

fbshipit-source-id: f48c7bb1709c0e1aba82f883bab9c4a9072bd288