DsfSubscriber: store switchIdToSystemPort instead of flat list
Summary:
FSDB sync carries switchIdToSystemPort info. But, DsfSubscriber puts that into
a flat list and processes it.

This worked fine when we had only single NPU in an RDSW. As we look to enable
multiple NPUs, we need to retain switchIdToSystemPort info and process it
appropriately.

This diff works towards that. Stacked diffs will use this data structure.

Reviewed By: msomasundaran

Differential Revision: D53167394

fbshipit-source-id: fb7fbbba0d9b78c8235f8bf598e7c19e5535dedc