Workaround for S391473: make noHostRoute optional to workaround 4.2 SAI bug
Summary:
In brcm-sai 4.2 there was a bug whereby they were noHostRoute=true for
link local neighbors unbeknownst to the NOS. This is not something a
sai SDK should ever do. The bug was recognized and fixed in brcm-sai
5.2. However, it seems that the old entries were never cleaned up.
4.2 was our SAI SDK ~2 years ago. Currently we are at 8.2

Recently, noHostRoute attribute that was added for a DSF use case
(diff stack: D52759880). Here we set the default for noHostRoute
to false (as per the SAI spec) and only set it to true for
DSF loopbacks. Warmboot tests for this went fine.

However, 9 devices showed a crash during warm boot with the
crash occurring during set of noHostRoute attribute from true to false.
https://fb.workplace.com/groups/fboss.push.investigations/posts/1758499181320290
On close inspection, it turns out these devices have not had a
cold boot/disruptive upgrade for ~2years.

https://fburl.com/ods/7erkodk0

On such devices if a LL neighbor that was created in 4.2 and has never
since flapped (update of such a neighbor would also fix the noHostRoute
setting since we would then recreate the neighbor -
https://fburl.com/code/qwtdl6o2). Then on WB, we will try to
update the attribute from true to false and crash.

The fix/workaround here makes noHostRotue a optional attribute. So
unless we explicitly set it - e.g. for the DSF loopback case. It will
be don't care and we will just retain the value set by brcm-sai sdk
If any.

Reviewed By: shri-khare

Differential Revision:
D53167586

Privacy Context Container: L1125642

fbshipit-source-id: 71b47d0d1b92c45975f8eedb8d8caa28dfcea206