Update profileID selection to consider masterLogicalPorts
Summary:
Current logic of `setPortToDefaultProfileIDMap` will consider all the ports within platform mapping and find a 'safe' profile for all the ports.

For example on J3, port 2 and 3 are in the same group. If 800G speed is needed, port 3 will be subsumed such that port 2 can bring up at 800G mode. However, the current logic will find 'safe' profile for port 2 and port 3, picking 400G for port 2 and 400G for port3. This is not desired for 800G testing.

To fix this, pass masterLogicalPorts into `setPortToDefaultProfileIDMap`. In the case of above port 2 and port 3 for J3:
1. If port 3 is not in the list of masterLogicalPorts, allow port2 to use 800G speed.
2. Else, assign 400G for port 2 and port 3.

Reviewed By: srikrishnagopu

Differential Revision: D52684250

fbshipit-source-id: 9833d96f71b55ed906d39432855084a018707b66