Introduce Serializable
Summary:
Introducing a base Serializable class to help us remove templated types whereever only encodablity is needed.
Unfortunately I've had to do some funkiness with the const correctness of PrimitiveNodes. Once this is all done, we might be able to clean up (maybe we can inline values for sets which are immutable)

Differential Revision: D49879838

fbshipit-source-id: a61cb34e2b9ad14789d1c01db4415905be1e8f30