set unresolved class id when nexthop is cpu port
Summary:
Routes that uses single nexthop and are unresolved points to cpu port as the nexthop to trigger neighbor resolution.

Based on the SAI spec, any route that points to CPU port will be treated as MYIP.

- For BCM, any /32 or /128 routes that points to CPU port will be treated as MYIP.
- For Tajo, any route (host route or subnet route) that points to CPU port will be treated as MYIP.

This causes the packet to land on mid pri queue which will cause BGP/Openr to flap.  In order to send these not MYIP routes to defauly queue, associate a class ID for those routes which will be matched against an ACL.
ACL will be programmed to send the packet to default queue.

Refer S390808 for more details.

We will be deploying this fix initially on Tajo platforms before enabling it on BCM though the tests will be enabled on both BCM and Tajo

Reviewed By: nivinl

Differential Revision: D53286428

fbshipit-source-id: 3987144ffa9178a0d773c3bb2aef943fc9f77108