define class id + gflag for unresolved route to cpu
Summary:
As titled, define a class ID + gflag for unresolved route to CPU.

Routes that uses single nexthop and are unresolved points to cpu port as the nexthop to trigger neighbor resolution. Associate the new class ID for those routes which will be matched against an ACL to send the packet to default queue.

Refer to S390808 for more details.

Reviewed By: nivinl

Differential Revision: D53286426

fbshipit-source-id: 539c5cc5b4452b0fb54651f0dec41dbf175072c9