Pass supportsAddRemovePort to genPortVlanCfg
Summary:
This is towards using the flag to determine if we use masterLogicalPorts to filter ports in the initial config.

The profileID selection logic was updated in D52684250, where the ports can pick max speed if the subsumed ports are not in masterLogicalPorts. However, TH with native SDK does not support add remove ports - meaning we'll need to keep the ports in switch state but in disabled mode.

Therefore, passing this flag into genPortVlanCfg and then later used in stacked diff.

Also refactor `onePortPerInterfaceConfig()` to take in swSwitch rather than several arguments (e.g. asic, loopback mode and addRemovePort from sw switch).

Reviewed By: msomasundaran

Differential Revision: D53338371

fbshipit-source-id: 930af7160ad19f5c8c15f78cc23e8f87792e11b9