Don't filter on masterLogicalPorts if platform does not support addRemovePort
Summary:
As titled. For platform that does not support addRemovePort, do not filter on masterLogicalPorts as it will remove subsumed ports that are not in materLogical Ports.

The profileID selection logic was updated in D52684250, where the ports can pick max speed if the subsumed ports are not in masterLogicalPorts. However, TH with native SDK does not support add remove ports - meaning we'll need to keep the ports in switch state but in disabled mode.

Reviewed By: harshitgulati18

Differential Revision:
D53340927

Privacy Context Container: L1125642

fbshipit-source-id: 3448ce19fc2596eb5d884aa58e134e08d0cb1459