Trap debug counter stat impl
Summary:
As titled. We should push for getting this
into the standard api. This is not hw specific.
Note that unlike other extension attributes (which we
can get/set) this is just a enum value. Hence the implementation
as a extension api. We adopted a similar approach for dramStats
before.

Reviewed By: shri-khare

Differential Revision: D53387892

fbshipit-source-id: 00dc415263b10ea04a73dbb9895f228d97cdadc9