Restrict link active state sync to dnx devices
Summary: In multi npu architecture, we sync link active state to swswitch when we establish connection to swswitch. This is enabled in hwtests also using a test thrift server. This is causing issues when hwtest is run on non dnx switches because we query sdk for txReadyStatuses with an empty port list. The fix is to do sync only if the switch supports fabric ports (covers voq and fabric switches)

Reviewed By: simuthus-fb

Differential Revision: D53589821

fbshipit-source-id: f0bb6463536da7f23cee4dbd86d24025270d4464