Update Sai switch delta processing sequence to support interface/vlan removal
Summary:
As titled. This is towards supporting flex ports from different vlans into one, as well as removing the other interface/vlan as a side effect.

(Previously we assume that the interfaces/vlans are static across warmboot, hence the removal sequence is not handled).

The new sequence (of related objects) are updated as followed -
1.  Remove routes
2. Remove neighbor entries from vlan (Arp, Ndp and Mac)
3. Remove neighbor entries from interfaces (Arp & NDP)
4. Removal local and remote interfaces
5. Add/Change local and remote interfaces
6. Add/Change neighbor entries from interfaces (Arp & NDP)
7. Add/Change neighbor entries from vlan (Arp, Ndp and Mac)
8. Add/Change routes

The sequence has become a bit complicated :) We should probably figure out the object dependency and document them (ideally removal and add/change should be a mirrored sequence)

Reviewed By: nivinl

Differential Revision:
D53031255

Privacy Context Container: L1125642

fbshipit-source-id: 20bec58ab6e967eaea8c651ca1a0366e32fee74c