Prevent removal of acl fb303 counters during acl reordering
Summary:
When Acls are repositioned because of addition/deletion of acls, changed acl entries can sometimes add acls followed by deletion of the same(in a different priority). This causes the fb303 counters to disappear and subsequently crash the box.

https://fburl.com/gdoc/96rz0n7q contains more details.

The fix involves adding a ref counter to acl stats and ensuring that acls are added/deleted only if the stats is >0/==0 respectively

Reviewed By: shri-khare

Differential Revision:
D53031584

Privacy Context Container: L1125642

fbshipit-source-id: 38a742c25d8fb3c1f7ca2ac96ab6be8e0a86362a