Towards building conditional attributes for debug counters
Summary:
Sai spec states {out|in}_reason_list as applicable to
out/in debug counter types respectively. Set/Get of even empty
reason list for inapplicable counter type gives unexpected
result.
E.g. sending a empty out reasons list of a in debug counter
resulted in a invalid param error
Get of out_reasons list on a in debug counter returned the
same reason list as for in reasons. This lead to a adapter
host key mismatch. Later leading to a crash.

We could fix this in all vendor SDKs or in our code. Conditional
attribute seems like the right strategy to fix this

Reviewed By: zechengh09

Differential Revision:
D53689204

Privacy Context Container: L1125642

fbshipit-source-id: 1f5bf7064ac2a27984ccfb25730b15e46e620f62