Fix thrift request error in HwSwitchThriftClientTable api
Summary: Swswitch has a thrift client table to connect to hwagents and execute any commands that need to be relayed from swswitch. Till now, we did not have any use for this table. Now we want to use this table in tests for collecting fabric reachability information from hwswitches. The previous code was using single evb which does not work with multiple hwagents. Also fixed an error in reconnecting channel which was causing bad returns from server

Reviewed By: zechengh09

Differential Revision: D53190200

fbshipit-source-id: c2b44a839d5c4cbcb37b4b4c5b99d6b71afd38e5