Start processing update when we have a valid switch state
Summary:
As titled. Imagine the mono initialization sequence

1. SwSwitch gets coldboot state from hw switch
2. SwSwitch notifies state observers on StateDelta(emptyState, coldbootState)
3. Enable lldp and start heartbeat thread
4. Transition to initialized state.

Note that between 2 and 4, state observers can process the state delta and schedule update on the state update thread - if that happens before step 4, the check of `isInitialized()` will fail.

Updating it to check if we have a valid switch state - if we do, we can schedule state update on it.

Reviewed By: msomasundaran

Differential Revision: D53729088

fbshipit-source-id: debe1e0671a42fb9ef1b7186aa080ae0e0f1bdc9