Fix updation of TLTimeSeries aggregate counters
Summary:
When multiswitch is enabled, the ods counters are exported from individual hwagents. However for backward compatability, the old counters are aggregated at swswitch and set in fb303 by swswitch as well. The setting of aggregated counter had a bug whereby the TLTimeSeries entries are updated "addValue" instead of setting new value. This means that the new value was added to the existing value. The value exported from SwSwitch is cumulative counter and hence cannot be used for addValue.

Fix:
Increment the counters using the delta between previous and new values

Reviewed By: jasmeetbagga

Differential Revision: D53341232

fbshipit-source-id: 9101a897cf321262b82269bf131fd92996b7338b