Fdr cell drops test, assert that we don't get spurious increments
Summary:
FDR cell drops in HW is a clear on read counter. So we
get a cumulative value by accumulating value obtained from HW.
In case our understanding was wrong and this was not a clear
on read counter, we would get a constant increment on every
updateStat after the first increment. So assert against this
in HW test

Reviewed By: zechengh09

Differential Revision: D53896188

fbshipit-source-id: 2c719d14d248b537abbdb7641323d026e6930a93