Add a API to check for no stats changing
Summary:
As titled. Will be used in several tests to
assert that we are not misaccumulating stats and getting constant
increments.
A common example is a cumulative stat interpreted as a clear on
read stat. If this happens then we get the same (cumulative) value
from HW but since we interpreted it as a clear on read, we add it
to existing value and get a constant increment

TODO: add cpu stats here, once D53795192 lands and bcm switch support this
api

Reviewed By: srikrishnagopu

Differential Revision:
D53896906

Privacy Context Container: L1125642

fbshipit-source-id: 8e2b016768314964b50a79d0a28900ae8142ea7a