Disable port removal and recreate for TAJO
Summary:
Current model for port speed change is to delete the existing port and recreate
with the new speed. However, TAJO has limitations on port delete and recreate
sequence with SDK1.42.8. There are some object dependencies which are not
guaranteed in SDK as of now and hence this functionality can only be used for
future releases SDK1.65.0 onwards.

To support port speed change with SDK1.42.8, removing the delete / recreate
flow, which will result in falling back to port speed change API being invoked to
achieve speed change. This however will work only for 100G - 200G where the
number of lanes are not changing. In cases like 200G - 400G or 400G - 100G,
this approach will not work and will need port delete/recreate flow to be
implemented.

These changes will not work standalone, need SDK changes from MT-676 to fix
an issue with port speed change from 200G -> 100G, this will be landed in a
different commit part of SDK drop for 1.42.8.6.

Reviewed By: srikrishnagopu

Differential Revision: D53713702

fbshipit-source-id: e8af9b83d1cd6f69d3975ea485996655e64ec40c