Read gflag instead of agent config for runMode
Summary: As titled. The purpose of having multi_switch in defaultCommandLine argument is also to set the gflag value. Changing this to 1. make sure value is correct even defaultCommandLine contains `multi_switch=false`, and 2. avoid the flag being reset in agent hw test, where new config can be applied and change the runMode.

Reviewed By: jasmeetbagga

Differential Revision: D53695985

fbshipit-source-id: f156c10718cf8e588824e4eca972c80b759961cf