Actually fail the setup_for_warmboot tests on failures
Summary:
When tests (like link tests) are run with --setup_for_warmboot flag, this is the tearDown path
1. tearDown
2. SwAgentInitializer::stopAgent
3. SwAgentInitializer::handleExitSignal
4. exit(0)

The process is unconditionally made to exit(0) ignoring any failures that the test has seen prior. This is causing all the cold_boot test runs to always pass.

Fixed it by doing exit(0) conditional on a flag, which is set to true for agent but false for tests.

Reviewed By: zechengh09

Differential Revision: D53897119

fbshipit-source-id: bcc05b4759a4c45ace4f721652ef6daa7b212efa