Refactor ACL entry configuration
Summary:
Today, an ACL entry is only created in HW if there is a trafficPolicy associated. The only exception is for DENY ACLs. This exception is also causing a trafficPolicy created for DENY ACLs not getting associated in the HW.
An additional issue found during this investigation was that the order of the ACL entries in switch config is not being honored.

This change refactors the entire logic. An ACL entry from config is always created irrespective of the presence of trafficPolicy. If a trafficPolicy is present, it will be associated.

Reviewed By: shri-khare

Differential Revision: D53198188

fbshipit-source-id: e86165a12bdbe81e1c5961940c1a3fe097fef0e2