Revised fan service config file for Tahan
Summary:
[FBOSS/Platform] Revised fan service config file for Tahan
This fan service config file works well with the revised sensor config file.
As the systen cooling / fan control scheme is not finalized, this config will
only decide the pwm value based on the simple table. (It will always run at
50% pwm)
Note that the actual PWM write does not happen (the path is incorrect intentionally,
as the cooling subsystem design is still under development)
Still, this config is more than enough to test the SW functionality and to be used
for Provisioning development and early DVT phase use.

Reviewed By: somasun

Differential Revision: D54131505

fbshipit-source-id: 6481db7fe60bdb7c29d8960cc194d6a26228720d