Add copp acl test to invariant copp helper
Summary: Issues with acl config are not being caught with the current prodInvariant test. This is because we are only verifying copp through packets which hit rxreasons. We also need to add a case to catch packets which will hit acls instead. This will ensure that if the acl is not programmed, the prodInvariant test will fail

Reviewed By: shri-khare

Differential Revision: D54204197

fbshipit-source-id: ea6e674d75f3bcbfccff5d75bef3d72589425081