Modify deleteUdfAclConfig
Summary:
With existing ACL config processing, an ACL is not installed without a traffic policy. After modifying that behavior to install ACLs without a traffic policy, the HwUdfTest.deleteUdfAclConfig started to fail because a stale ACL entry remained after the test.

This change makes sure the ACL entry is cleaned up appropriately.

Differential Revision: D54278503

fbshipit-source-id: 46af8f52c83a3671f8461a8fff917ee774393fa2