Improve error handling and logging
Summary:
If a PCI SubDevice already exists, ioctl will fail (ret != 0), this is not
necessarily a failure condition, and we don't have to report error and abort.
This diff changes this path to error out only if the actual device does not
exist irrespective of the ioctl return code. We just log the ioctl return code.

Reviewed By: somasun

Differential Revision:
D54361583

Privacy Context Container: L1125642

fbshipit-source-id: ef32e1801439790dd84dc7ce3177bf1abab5ead1