Dynamically find the VDM config and data offsets
Summary:
The CMIS spec does not specify the fixed location for a given VDM config and values. We need to scan the VDM config pages (0x20-0x23) and find the corresponding location in pages 0x24-0x27 where these VDM data will be stored.

We are scanning the VDM config locations and storing the page, regOffset and length values in CmisModule. The scan will be done just once.

Reviewed By: harshitgulati18

Differential Revision:
D54282206

Privacy Context Container: L1125642

fbshipit-source-id: 207a54a7de8945a241c0622a9a4df6b189f60876