Export all sensor values with sensor_read prefix
Summary:
The sensor values are currently exported without any prefix, this
makes it harder to set up monitoring configs for them. A catch all regex (.*)
is bad, as there are gazillion other counters which will get exported. In this
diff, we export sensor values with sensor_read prefix (which is used for
failure count as well).

Slowly, we will phase out the counters which are currently exported without any
prefix.

Reviewed By: jasmeetbagga

Differential Revision:
D54447325

Privacy Context Container: L1125642

fbshipit-source-id: 996a9ad48c1ae9e23d21ebbf80bac37d2b242531