Refactor: create TransceiverConfig in Mgr, pass it to Transceivers
Summary:
To avoid getting the transceiver config from manager (transceiverConfigOverrides), we can pass that directly to the module (as a TransceiverConfig), and avoid referencing the manager in the Transceiver.

There is a single copy of this config (in TransceiverManager) since its common to all transceivers, and The Transceiver instances will get a shared pointer.

Reviewed By: harshitgulati18

Differential Revision: D53845550

fbshipit-source-id: 2a26db184423db27ae390377d6bd73ba7e442a81