Refactor ACL entry configuration
Summary:
Today, an ACL entry is only created in HW if there is a trafficPolicy associated. The only exception is for DENY ACLs. This exception is also causing a trafficPolicy created for DENY ACLs not getting associated in the HW.
An additional issue found during this investigation was that the order of the ACL entries in switch config is not being honored.

This change refactors the entire logic. An ACL entry from config is always created irrespective of the presence of trafficPolicy. If a trafficPolicy is present, it will be associated.

Additional context:
This diff is a copy of the original diffs committed via D53198188. No changes are made to the stack from the original.
This stack was previously reverted using D54235546. The reason was failing tests in conveyor. As there was a immediate requirement for conveyor to continue, the diffs were reverted.

After further analysis, the failures were observed mostly in the new tests added in this stack. These tests are disabled with reasons described in  D54320486.

1 existing test was also failing, deleteUdfAclConfig. This failure was not caught on-diff because of a failing signal at the time of landing. This test is also fixed in the first commit of this stack.(D54278503)

A more reserved approach will be taken for landing this stack.
The first 3 diffs will be landed first since the only change here is they allow ACl creation without traffic policy.
The last 2 diffs will then be landed. The DENY ACL priorities are fixed in those 2 diffs.

Reviewed By: shri-khare

Differential Revision: D54286548

fbshipit-source-id: cf388a839ab216e4cc37910628d5130d6cad3602