Adjust the PFC WD detection/recovery interval based on trigger mechanism used
Summary:
When PFC WD tests are done, if traffic is used to trigger deadlock detection, we
need to use lowest possible intervals. However, if we are programing registers to
trigger PFC frames, then we need to keep the detection interval really high to
avoid flood of detection events. Make this differentiation based on asics.

Reviewed By: jasmeetbagga

Differential Revision: D54622453

fbshipit-source-id: 6c794dd6029798c1ec05abfd87f22316459ae744