Refactor: Moving FirmwareUpgrader and improve interfaces it uses [1 of 5]
Summary:
This diff adds the basic functions required to use TransceiverImpl instead of TransceiverI2CApi in FirmwareUpgrader. This also helps us avoid referencing TransceiverManager in CmisModule.

The CmisModule passes a TransceiverI2CApi to the FirmwareUpgrader Object, when it could simply pass the TransceiverImpl object instead (TransceiverImpl is already passed to the CmisModule at construction).

Reviewed By: harshitgulati18

Differential Revision: D53876849

fbshipit-source-id: 43e9d7cc6ec0b016bc40a660727545e6953ed8af