Refactor: Moving FirmwareUpgrader and improve interfaces it uses [4 of 5]
Summary:
Add ability to specify delay in post I2C Write in TransceiverImpl. The upgrade firmware classes will use the TransceiverImpl for direct I2C writes. The FirmwareUpgrader will take care of adding the appropriate delay when writing to the Transceiver. The main goal of this is to replace TransceiverI2CApi with TransceiverImpl.

This helps us avoid the extra delay currently hardcoded in TransceiverImpl, to avoid long upgrade time for FW.

If we enable delay also on the TransceiverImpl, this will increase the overall upgrade time: assuming ~600KB, we write 32 bytes at a time and then add the delay (20ms in TransceiverImpl and another 5ms in CdbCommandBlock). The total delay time will cause upgrade time to be around ~7min.

Reviewed By: harshitgulati18

Differential Revision: D54094401

fbshipit-source-id: 68a0661f21c338e0c59bd422c8427c1d35cee968