Store cumulative fb303 value in HwFb303Stats map
Summary: HwFb303Stats uses fb303 monotonous counters to store and export values. These counters do not provide a method to get the cumulative value. Currently we read the values by looking it up in fb303 statsmap using string name. This takes lock to synchronize between all threads that might increment the counter. However the HwFb303 stats value is updated only from stats thread and hence another thread locking is not really needed to get the cumulative value. Store the cumulative value in the map along with monotonous counter so that it can be accessed from the stats thread

Reviewed By: jasmeetbagga

Differential Revision: D54600436

fbshipit-source-id: 566ae36c2a073af3fd5ce3973c9631302578b526