Ignore empty stats update callbacks
Summary: The current port stat verification tests are getting empty stats sometimes and using retry to succeed. On debugging this further, it was noticed that we are getting empty stats callback on swswitch thrift handler sometimes with regular update. These empty updates are not pushed by hwagent based on debugs. One suspicion is that large stats message might be triggering it since i don't see this behavior for other sinks such as link events . I will experiment with msg size and follow up with thrift team on this. Meanwhile ignoring the empty updates

Reviewed By: jasmeetbagga

Differential Revision: D54708664

fbshipit-source-id: c811fa45851a20b1874554a5281bfe17cbafa719