Always load upstream kmods
Summary:
`processKmods` is not always run on PM startup depending on whether or not we want to reload/install all of the kmods. Since the upstream kmods need to be loaded no matter what, it is better to load them regardless of how PM was started.

This moves loading upstream kmods to a separate function which will always be executed at the beginning of PM startup.

Reviewed By: somasun

Differential Revision: D54779917

fbshipit-source-id: 943fb66a161429cd8bb3a1b8f279f4f3868ff783