remove stats snapshot
Summary: This module has not been used in years. Pruning it and some tests to make oss build easier. We can probably remove RocksDb altogether but for now I'm keeping it around as we might be able to reuse some of the code for initializing the db.

Reviewed By: srikrishnagopu

Differential Revision:
D54497674

Privacy Context Container: L1158612

fbshipit-source-id: 62c0a2f4b93ec8f4fd20b5d7880245e6f27ad37b