Fix empty stats update happening in tests
Summary: The empty stats updates seen in test was happening because the mono stats fix up was running in tests in multi mode also because the multi switch default cmd line flag in agent config is not set in multi switch tests. Instead it sets FLAGS_multi_switch directly. Removing the check for empty stats update and using FLAGS_multi_switch in stats update to prevent the overwrite

Reviewed By: jasmeetbagga

Differential Revision: D54783217

fbshipit-source-id: 1ca78bc4446354dd16abd2e0f72842a8e93175f3