Allow for passing a friendly name to stream log
Summary: Switch agent logs are annoying to deal with right now because we only log the ip addr for remote fsdb hosts. Would be nice to get the actual device name. Adding an option to pass device name and use that to log if available

Reviewed By: wilsonwinhi

Differential Revision:
D54861776

Privacy Context Container: L1158612

fbshipit-source-id: 747ba0459fa4c71544261b2ba752b02c47d36b0a