Fix qos policy validation error in mono test
Summary:
One mono test fails on state validation with error
```
Both DSCP to TC and TC to Queue maps must be provided in valid qos policies
```
Adding TC to queue map to the test config. This was not seen with HwTest because the validation is not excercised. Also this was not seen with multi because multi does validation only in swswitch and not in hwswitch.

TODO - enable state validation in hwswitch with multi

Reviewed By: srikrishnagopu

Differential Revision: D54936373

fbshipit-source-id: b6d219e856795267d8116f82ff93648816a0be58