Provide initialSync cb when subscribed leaf is not present
Summary:
Issue being addressed: Initial sync is never completed by fsdb server when subscription is for a leaf that is not published by the publisher.

Cause: When a publisher root connects and a subscription is initially served, if the initial chunk to be served has no value, fsdb server was not offering any response to the subscription stream.

Fix for server side issue: Initial response must be offered by FSDB server to a connected subscription stream. Specifically, for this case serve a heartbeat message to the stream.

Fix for client side behavior: Client can register a SubscriptionStateChangeCb to be notified when subscription stream connects/disconnects. The callback is invoked even when initial sync results in empty data.

Reviewed By: peygar

Differential Revision: D54819333

fbshipit-source-id: 2491b5d7de48b0c95721d85a1cfec689f711bf50