Define gflag and class ID for connected subnet routes
Summary:
This change came about after investigating the Dark Isilon design.
Cisco/TAJO will not be supported until migration to 1.65 is done.

Unresolved traffic to interface subnet routes are sent to low prio queue. In the presence of a DENY data ACL, this traffic will be dropped.
On BRCM, the packet would be dropped, while TAJO allows the packet irrespective of the presence of ACL.

Note: Host routes are unaffected by a DROP ACL. This is because an rx reason entry is installed which will make the drop action is overriden.

This subnet route will modify the route metadata and then match on a higher priority ACL entry (compared to the DENY ACL) with this new route metadata value.

This can eventually become the default for RSWs but gated with a flag for now.

Dependent on D54320534.

Reviewed By: srikrishnagopu, msomasundaran

Differential Revision: D53831334

fbshipit-source-id: 0020c07ca621915a69c59b35c79946eddbb53ade